pyproject.toml
[tool.poetry]
name = "test-task-middle"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"
packages = [{include = "test_task_middle"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
asyncpg = "^0.28.0"
sqlalchemy = "^2.0.22"
alembic = "^1.12.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
fastapi = "^0.104.0"
numpy = "^1.26.1"
torch = ">=2.0.0, !=2.0.1, !=2.1.0"
pandas = "^2.1.1"
email-validator = "^2.0.0.post2"
uvicorn = "^0.23.2"
python-jose = "^3.3.0"
passlib = "^1.7.4"
matplotlib = "^3.8.0"
scikit-learn = "^1.3.2"
statsmodels = "^0.14.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
coverage = "^7.3.2"
httpx = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
